requirements.txt
torch==2.4.1 -f https://download.pytorch.org/whl/cu124
torchvision==0.19.1 -f https://download.pytorch.org/whl/cu124
torchaudio==2.4.1 -f https://download.pytorch.org/whl/cu124
soundfile==0.12.1
tqdm==4.67.0
librosa==0.9.2
numpy==2.0.2
PyYAML==6.0.2
ml_collections==1.0.0
